+++ addOns/scripts/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- NPE when using some scripts after re-installing the scripts add-on.
 
 ## [45.9.0] - 2025-03-25
 ### Fixed